Fix the three pre-existing clippy lints
Clears the advisory warnings CI has been printing since before the rename,
so `cargo clippy --all-targets -- -D warnings` is now clean:

- icon.rs: drop a u32 -> u32 no-op cast
- config.rs: build the test Configs via struct update syntax instead of
  assigning fields after Default::default()

No behaviour change; 14/14 tests still pass.

+++ src/icon.rs
@@ -82,7 +82,7 @@ pub fn rgba(size: u32) -> Vec<u8> {
 
     // Box downsample ss×ss -> size×size.
     let mut out = vec![0u8; (size * size * 4) as usize];
-    let cnt = (ss * ss) as u32;
+    let cnt = ss * ss;
     for y in 0..size {
         for x in 0..size {
             let (mut r, mut g, mut b, mut a) = (0u32, 0u32, 0u32, 0u32);
